+++ frontend/components/thematicMap/ThematicMap.tsx
@@ -1,59 +1,10 @@
+import GridBox from '@components/gridBox/GridBox';
 import { FieldDataType } from '@interfaces/fieldData.interface';
 import { RangesType } from '@interfaces/range.interface';
 import { getAllFieldsData, getFieldData } from '@services/data.service';
-import { convertStrToInt } from '@utils/strToInt';
+
 import ranges from 'data/ranges';
 import { useEffect, useState } from 'react';
-
-import { BsMoisture } from 'react-icons/bs';
-import { TbTemperaturePlus } from 'react-icons/tb';
-import { WiHumidity } from 'react-icons/wi';
-
-const GridBox = ({
-  param,
-  type,
-  range,
-}: {
-  param: FieldDataType;
-  type: 'moisture' | 'temperature' | 'humidity';
-  range?: RangesType;
-}) => {
-  const created_at = new Date(param.created_at);
-  const value = convertStrToInt(param.value);
-  const block = param.block;
-
-  // filter out the range that the value falls into and get the color
-  let color: [number, number, number] = [0, 0, 0];
-
-  if (range) {
-    if (value < range.level_1.value) {
-      color = range.level_1.color;
-    } else if (value >= range.level_1.value && value < range.level_2.value) {
-      color = range.level_2.color;
-    } else {
-      color = range.level_3.color;
-    }
-  }
-
-  return (
-    <div
-      className={
-        'h-44 w-44 flex flex-col gap-3 justify-center items-center text-2xl text-white font-semibold'
-      }
-      style={{
-        backgroundColor: `rgb(${color[0]}, ${color[1]}, ${color[2]})`,
-      }}
-    >
-      <div className="text-3xl">
-        {type === 'moisture' && <BsMoisture />}
-        {type === 'temperature' && <TbTemperaturePlus />}
-        {type === 'humidity' && <WiHumidity />}
-      </div>
-
-      <div>{`${value}%`}</div>
-    </div>
-  );
-};
 
 const TimestampView = ({ time }: { time: Date }) => {
   return (
